+++ cmd/entire/cli/root.go
@@ -83,6 +83,7 @@ func NewRootCmd() *cobra.Command {
 	cmd.AddCommand(newResumeCmd())
 	cmd.AddCommand(newCleanCmd())
 	cmd.AddCommand(newResetCmd())
+	cmd.AddCommand(newStopCmd())
 	cmd.AddCommand(newSetupCmd())
 	cmd.AddCommand(newEnableCmd())
 	cmd.AddCommand(newDisableCmd())